(refactor) Transition material design components to default.jsx versus index.jsx
